session-9-authN-authZ-validation/hw/requirements.txt
# Flask web framework and extensions
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# Authentication and Security
PyJWT>=2.0.0
Werkzeug>=2.0.0

# HTTP client for testing
requests>=2.25.0

# Type hints (Python < 3.9 compatibility)
typing-extensions>=4.0.0
